+++ plugin-gradle/src/test/java/com/diffplug/gradle/spotless/DiffMessageFormatterTest.java
@@ -393,4 +393,69 @@ public class DiffMessageFormatterTest extends ResourceHarness {
 				"        -47␍␊",
 				"    ... (1952 more lines that didn't fit)");
 	}
+
+	@Test
+	public void oneMoreLineThatDidntFit() throws Exception {
+		// com.diffplug.spotless.extra.integration.DiffMessageFormatter.MAX_CHECK_MESSAGE_LINES
+		// defaults to 50, so we create a diff that would be exactly 50 lines long
+		// The test is to ensure diff does not contain "1 more lines that didn't fit"
+		StringBuilder builder = new StringBuilder();
+		for (int i = 0; i < 25; ++i) {
+			builder.append(i);
+			builder.append(i < 1 ? "\n" : "\r\n");
+		}
+		SpotlessTask task = create(setFile("testFile").toContent(builder.toString()));
+		assertTaskFailure(task,
+				"    testFile",
+				"        @@ -1,25 +1,25 @@",
+				"         0",
+				"        -1␍␊",
+				"        -2␍␊",
+				"        -3␍␊",
+				"        -4␍␊",
+				"        -5␍␊",
+				"        -6␍␊",
+				"        -7␍␊",
+				"        -8␍␊",
+				"        -9␍␊",
+				"        -10␍␊",
+				"        -11␍␊",
+				"        -12␍␊",
+				"        -13␍␊",
+				"        -14␍␊",
+				"        -15␍␊",
+				"        -16␍␊",
+				"        -17␍␊",
+				"        -18␍␊",
+				"        -19␍␊",
+				"        -20␍␊",
+				"        -21␍␊",
+				"        -22␍␊",
+				"        -23␍␊",
+				"        -24␍␊",
+				"        +1␊",
+				"        +2␊",
+				"        +3␊",
+				"        +4␊",
+				"        +5␊",
+				"        +6␊",
+				"        +7␊",
+				"        +8␊",
+				"        +9␊",
+				"        +10␊",
+				"        +11␊",
+				"        +12␊",
+				"        +13␊",
+				"        +14␊",
+				"        +15␊",
+				"        +16␊",
+				"        +17␊",
+				"        +18␊",
+				"        +19␊",
+				"        +20␊",
+				"        +21␊",
+				"        +22␊",
+				"        +23␊",
+				"        +24␊");
+	}
 }